Cities List, Header ID Change
Added a list of the 3 largest cities in each state (Source: Wikipedia), and changed the ID of the header column to be more general. In some cities, the original ID of the title header is component_3, and in some it is component_4, so the ID to search for was changed to the parent div instead.